Add files via upload
The technology of chat application has been available for years but adoption has only recently taken place. This project is an example of a chat server. It is made up of two applications - the client application, which runs on the user's web browser and the server application, running on any network servers.